feat: hoist shared icon
